+++ libosip2-2.2.2/src/osipparser2/osip_message_to_str.c
@@ -1025,7 +1025,7 @@ _osip_message_to_str (osip_message_t * sip, char **dest,
     char tmp2[15];
 
     total_length += size;
-    sprintf (tmp2, "%i", size);
+    sprintf (tmp2, "%i", (int)size);
     /* do not use osip_strncpy here! */
     strncpy (content_length_to_modify + 5 - strlen (tmp2), tmp2, strlen (tmp2));
   }
